requirements.txt
absl-py==1.4.0
accelerate==1.0.1
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
alembic==1.10.4
anndata==0.9.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argcomplete==3.0.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bioc==2.1
bleach==6.0.0
boto==2.49.0
braceexpand==0.1.7
bypy==1.8.2
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmaes==0.9.1
cmake==3.25.0
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.7
cramjam==2.6.2
crcmod==1.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.6
docopt==0.6.2
efficientnet-pytorch==0.7.1
einops==0.6.0
executing==1.2.0
fasteners==0.18
fastjsonschema==2.16.3
fastparquet==2023.2.0
filelock==3.9.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.10.0
ftfy==6.1.1
gcs-oauth2-boto-plugin==3.0
get-annotations==0.1.2
google-apitools==0.5.32
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-reauth==0.1.1
greenlet==2.0.2
grpcio==1.53.0
gsutil==5.23
h5py==3.11.0
httplib2==0.20.4
huggingface-hub==0.26.2
idna==3.4
imageio==2.28.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
install==1.3.5
intervaltree==3.1.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==4.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.16
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.11.8
lit==15.0.7
llvmlite==0.41.1
lxml==5.1.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
monai==1.3.2
monotonic==1.6
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
munch==2.5.0
natsort==8.4.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
negbio==0.9.4
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.58.1
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.16.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
optuna==3.1.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
polars==0.17.0
pretrainedmodels==0.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pynndescent==0.5.13
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytorch-fid==0.3.0
pytorch-lightning==1.9.0
pytorch-msssim==1.0.0
pytz==2022.7.1
pyu2f==0.1.5
pyvips==2.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retry-decorator==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
safetensors==0.4.5
scanpy==1.9.8
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.13.2
# Editable install with no version control (segmentation-models-pytorch==0.3.2)
-e /storage/ICRGW/ICRGW/packages/segmentation_models.pytorch-master
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
session-info==1.0.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.13
stack-data==0.6.2
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==1.0.11
tinycss2==1.2.1
tokenizers==0.13.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==0.11.4
torchvision==0.15.1+cu118
# Editable install with no version control (torchxrayvision==1.1.1)
-e /storage/ScientificPrograms/Diffusion/code/torchxrayvision-master
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
umap-learn==0.5.7
uri-template==1.2.0
urllib3==1.26.13
wcwidth==0.2.6
webcolors==1.12
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
yarl==1.9.2
zipp==3.15.0
